requirements.txt
# ####################### #
# Created for Python 3.10 #
# ####################### #

# ##################### #
#  install torch with   #
# specific cuda version #
# ##################### #
torch==2.1.* --index-url https://download.pytorch.org/whl/cu121
torchvision==0.16.* --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.1.* --index-url https://download.pytorch.org/whl/cu121

# ######################## #
# regular pip dependencies #
# ######################## #
av
black
coverage[toml]
cython==3.0.*
easydict
future
gdown
h5py
imageio
imagesize
jinja2
matplotlib
# # munkres is a package for the hungarian (matching) algorithm
munkres
natsort
ninja
numpy
opencv-python
Pillow
pycocotools==2.0.*
pylint
pyyaml==6.0.*
scipy
# # sphinx is used to build the docs
sphinx
# # use tb nightly, due to torchreid installing it and it bein incompatible with the regular tensorboard
tb-nightly
torcheval
tqdm
visdom


# ################## #
# local dependencies #
# ################## #

# Set up halpecocotools
# fixme do we still need halpecocotools without AlphaPose ?
-e dependencies/halpecocotools/PythonAPI/
# Torchreid - install requirements and Set up torchreid and cythonize
# somehow installing requirements does not work just by calling setup
-e dependencies/torchreid/
# install PoseTrack21 evaluation toolkit
-e dependencies/PoseTrack21/eval/posetrack21/
# Alpha Pose plus its requirements - make sure this will be installed last!
# -e dependencies/AlphaPose_Fork

# own project
-e .